+++ dev/app/server.R
@@ -5,7 +5,7 @@ my_server <- function(input, output, session) {
     mod_eda_freq_table_server(id = "grouping", r_data = r_mtcars)
 
     # --- Select ---
-    mod_eda_select_server(id = "test", r_data = r_mtcars)
+    mod_select_server(id = "test", r_data = r_mtcars)
 }
 
 my_server